Moved tables into correct apps
