+++ django_project/catalogue/utils.py
